+++ modules/core/src/main/java/org/apache/ignite/plugin/extensions/communication/MessageWriter.java
@@ -46,11 +46,11 @@ public interface MessageWriter {
 
     /**
      * Writes message header.
-     *
      * @param type Message type.
      * @param fieldCnt Fields count.ß
+     * @return {@code true} if successfully. Otherwise returns {@code false}.
      */
-    public boolean writeHeader(byte type, byte fieldCnt);
+    public boolean writeHeader(short type, byte fieldCnt);
 
     /**
      * Writes {@code byte} value.
