+++ modules/core/src/test/java/org/apache/ignite/loadtests/communication/GridTestMessage.java
@@ -127,7 +127,7 @@ class GridTestMessage implements Message, Externalizable {
     }
 
     /** {@inheritDoc} */
-    @Override public byte directType() {
+    @Override public short directType() {
         return 0;
     }
